Avoid reading out of bounds for strings
Rust's str is not null terminated. So simply casting the pointer like
let xpath = xpath.as_ptr() as *const i8;
will result in a pointer to a not null terminated string and
can thus not properly used when interfacing C libraries.

The preferred alternative is to use CStr or CString that makes
sure that the string is null terminated.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>